build: make fonts immutable
The generated subfonts contain a hash in the filename, so it makes sense to override Netlify's
default behaviour of requiring a revalidation since there's a different path if the asset changes.

The Netlify [`_headers`] file appears not to allow for partial matching, so the assets in the route
directly need moving to a dedicated subpath first.

[`_headers`]: https://docs.netlify.com/routing/headers/