fix(build): bypass Husky install step when not installed
The `prepare` script is still run when using the `--production` flag, so Husky was not installed
but still called, leading to failures. This checks that Husky is installed before running it.

It also adds the `name` property to `package.json`. This is optional, but due to a
[bug in npm-lifecycle] it has to be included. This will be resolved when moving to npm 7, which
uses [`@npmcli/run-script`] instead.

These were both caused by the upgrade of Husky (6e0c4e005214fcaaafc540a724ea8d844618dc49).

[Bug in npm-lifecycle]: https://stackoverflow.com/a/60112825
[`@npmcli/run-script`]: https://www.npmjs.com/package/@npmcli/run-script