fix(build): don't use Parcel's built-in server
HTML files other than `index.html` have never worked properly with `parcel serve`, as in they could
not be accessed. Turns out that it's designed solely for [SPAs], which is a strange choice and
is incompatible with this 'traditional' site.

This changes Parcel usage to only `watch`, and instead uses the [serve] as the server.

The port changes from `1234` to `5000` (serve's default); the former is still used by
`parcel watch` to provide the websocket for Hot Module Replacement.

[serve]: https://www.npmjs.com/package/serve
[SPAs]: https://github.com/parcel-bundler/parcel/issues/1354#issuecomment-388604971